Part 1 gone mad
So I made a program for part 1 of today which is completely wrong: 600.
Because I was frustrated with the wrong answer I decided to use the output for my friend's input as a sort of ratio with his actual output - 1820/525. I then multiplied my answer by this, and lo and behold: a correct answer! (600*(1820/525) = 2080). This is super weird as it didn't work with any of my other friend's inputs lol.